Fix the value of hl-sexp-overlay when the buffer is cloned with make-indirect-buffer.
make-indirect-buffer copy the overlay referenced by hl-sexp-overlay
but use the newly created buffer.

This is annoying because hl-sexp-overlay isn't change, so moving in
the new buffer move the overlay in the original buffer.

To handle this situation I've added a function that does what
make-indirect-buffer did -- copy + update buffer -- and set this
overlay to hl-sexp-overlay.

This function also deletes the overlay created by
make-indirect-buffer.